Распределил компоненты по модулям. eslintignore.
