+++ src/common.js
@@ -1,0 +1,8 @@
+export const AddedComponentPosition = {
+  BEFORE_BEGIN: `beforebegin`,
+  AFTER_BEGIN: `afterbegin`,
+  BEFORE_END: `beforeend`,
+  AFTER_END: `afterend`
+};
+
+export const render = (container, template, place) => container.insertAdjacentHTML(place, template);
