+++ src/view/trip-information.js
@@ -1,0 +1,22 @@
+import {render, AddedComponentPosition} from "../common.js";
+import {createTripPriceTemplate} from "./trip-price.js";
+
+const createTripInformationContainerTemplate = () =>
+  `<section class="trip-main__trip-info  trip-info">
+    <div class="trip-info__main">
+      <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
+
+      <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;20</p>
+    </div>
+  </section>`;
+
+export const createTripInformationTemplate = () => {
+  // Маршрут
+  const pageBodyElement = document.querySelector(`.page-body`);
+  const tripMainElement = pageBodyElement.querySelector(`.trip-main`);
+  render(tripMainElement, createTripInformationContainerTemplate(), AddedComponentPosition.AFTER_BEGIN);
+
+  // и стоимость
+  const priceAndRouteElement = tripMainElement.querySelector(`.trip-main__trip-info`);
+  render(priceAndRouteElement, createTripPriceTemplate(), AddedComponentPosition.BEFORE_END);
+};
