+++ src/view/trip-price.js
@@ -1,0 +1,4 @@
+export const createTripPriceTemplate = () =>
+  `<p class="trip-info__cost">
+    Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+  </p>`;
